+++ src/components/select/select.test.jsx
@@ -1,6 +1,5 @@
 import React, { createRef } from 'react';
 import {
-  fireEvent,
   render,
   screen,
 } from '@testing-library/react';
@@ -11,8 +10,8 @@ import Select from './select.jsx';
 describe('src/components/select/select', () => {
   const baseListOptions = ['foo', 'bar'];
   const baseListOptionRefs = baseListOptions.map(() => createRef());
-  const baseListOptionElements = baseListOptions.map((option, index) => {
-    return (<ListOption key={option} ref={baseListOptionRefs[index]} value={option}>{option}</ListOption>);
+  const baseListOptionElements = ({ onSelect }) => baseListOptions.map((option, index) => {
+    return (<ListOption key={option} onSelect={onSelect} ref={baseListOptionRefs[index]} value={option}>{option}</ListOption>);
   });
   const requiredProps = {
     id: 'test-id',
@@ -45,7 +44,7 @@ describe('src/components/select/select', () => {
       expect(screen.queryByText('foo')).not.toBeInTheDocument();
       expect(screen.queryByText('bar')).not.toBeInTheDocument();
 
-      fireEvent.keyDown(document.activeElement, { key: 'Enter' });
+      userEvent.keyboard('{Enter}');
       expect(screen.getByText('foo')).toBeInTheDocument();
       expect(screen.getByText('bar')).toBeInTheDocument();
     });
@@ -56,7 +55,7 @@ describe('src/components/select/select', () => {
       expect(screen.getByText('foo')).toBeInTheDocument();
       expect(screen.getByText('bar')).toBeInTheDocument();
 
-      fireEvent.keyDown(document.activeElement, { key: 'Escape' });
+      userEvent.keyboard('{Escape}');
       expect(screen.queryByText('foo')).not.toBeInTheDocument();
       expect(screen.queryByText('bar')).not.toBeInTheDocument();
     });
@@ -73,7 +72,7 @@ describe('src/components/select/select', () => {
       render(<Select {...requiredProps}>{baseListOptionElements}</Select>);
       userEvent.click(screen.getByLabelText('Test label'));
       userEvent.tab();
-      fireEvent.keyDown(document.activeElement, { key: 'ArrowDown' });
+      userEvent.keyboard('{ArrowDown}');
 
       expect(document.activeElement.innerHTML).toBe('bar');
     });
@@ -211,7 +210,7 @@ describe('src/components/select/select', () => {
       userEvent.click(screen.getByLabelText('Test label'));
       userEvent.click(screen.getByText('foo'));
       userEvent.click(screen.getByLabelText('Test label'));
-      fireEvent.change(document.activeElement, { target: { value: '' } });
+      userEvent.clear(document.activeElement);
 
       expect(document.activeElement).toHaveValue('');
     });
@@ -301,25 +300,28 @@ describe('src/components/select/select', () => {
       expect(screen.getByText('foo')).toHaveAttribute('aria-selected', 'true');
     });
 
-    it('keeps the selected value selected even when value type is complex', () => {
+    it('keeps the selected value selected even when value type is complex', async () => {
       const listOptions = [{ id: 0, label: 'foo' }];
       const listOptionRefs = listOptions.map(() => React.createRef());
-      const listOptionElements = listOptions.map((option, index) => {
-        return (<ListOption key={option.id} ref={listOptionRefs[index]} value={option}>{option.label}</ListOption>);
+      const listOptionElements = ({ onSelect }) => listOptions.map((option, index) => {
+        return (<ListOption key={option.id} onSelect={onSelect} ref={listOptionRefs[index]} value={option}>
+          {option.label}
+        </ListOption>);
       });
 
       render(
         <Select
           {...requiredProps}
+          listOptionRefs={listOptionRefs}
           value={listOptions[0]}
           displayValueEvaluator={value => value.label}
         >
           {listOptionElements}
         </Select>);
       userEvent.click(screen.getByLabelText('Test label'));
+      expect(screen.getByText('foo')).toHaveAttribute('aria-selected', 'true');
       userEvent.click(screen.getByText('foo'));
-      userEvent.click(screen.getAllByLabelText('Test label')[0]);
-
+      userEvent.click(screen.getByLabelText('Test label'));
       expect(screen.getByText('foo')).toHaveAttribute('aria-selected', 'true');
     });
   });
@@ -363,7 +365,7 @@ describe('src/components/select/select', () => {
       userEvent.click(screen.getByLabelText('Test label', { selector: 'input' }));
       userEvent.tab();
       expect(screen.getByRole('button', { name: 'Remove selected choice' })).toHaveFocus();
-      fireEvent.keyDown(screen.getByRole('button', { name: 'Remove selected choice' }).firstChild, { key: 'Enter', code: 'Enter' });
+      userEvent.keyboard('{Enter}');
       expect(screen.getAllByLabelText('Test label')[0]).toHaveValue('');
     });
 
@@ -402,7 +404,7 @@ describe('src/components/select/select', () => {
 
       expect(changeValue).toEqual('foo');
 
-      fireEvent.keyDown(screen.getByRole('button', { name: 'Remove selected choice' }).firstChild, { key: 'Enter', code: 'Enter' });
+      userEvent.click(screen.getByRole('button', { name: 'Remove selected choice' }));
       userEvent.click(screen.getByLabelText('Test label'));
       userEvent.click(screen.getByText('bar'));